requirements.txt
qdrant-client
transformers
sentence-transformers
torch
fastapi
uvicorn
pyyaml
